api/requirements.txt
antlr4-python3-runtime==4.10
anyio==3.5.0
asgiref==3.5.0
click==8.1.2
greenlet==1.1.2
h11==0.13.0
idna==3.3
kwonly-args==1.0.10
mccabe==0.6.1
nodeenv==1.6.0
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
pyright==1.1.235
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
typing_extensions==4.1.1
fastapi==0.75.1
uvicorn==0.17.6
json-cfg==0.4.2
flake8==4.0.1
pyright==1.1.235
SQLAlchemy==1.4.35
sqlalchemy-views==0.3.1
psycopg2-binary==2.9.3
urllib3==1.26.9
autopep8==1.6.0
antlr4-python3-runtime==4.10
PyJWT==2.6.0
python-dotenv==1.0.1
gunicorn==21.2.0
boto3
python-dotenv
openpyxl
cryptography==41.0.7
requests




